ahjo: Add Ahjo XML HTTP/ZIP interface
Implement scanning and reading of the ZIP files from the HTTP interface
of Ahjo XML API.  Also add a stub for an importer that imports the
documents from this API to the database.

The scanner is in scanner.py.  It scans and parses the directory
contents from the HTML sent by the server and then returns DocumentInfo
objects for each ZIP file found from the server.  These document infos
allow opening the XML file from inside the ZIP file and then parsing it
with the AhjoDocument class.

The database importer is in db_importer.py.  The importer needs a new
database model ImportedFile, which tracks the names and modification
times of the already imported files.  Add that model and a database
migration for it.

Also add some tests for the HTML directory list parsing.

Some of the code is inspired by OpenAhjo and also the ignore lists are
copied from it.  The OpenAhjo files used as a reference are scanner.py
and doc.py in ahjodoc directory at
https://github.com/City-of-Helsinki/openahjo/tree/614dc96e91762f762a6e60977251e9ab270ef192